requirements.txt
aiogram
motor
python-dotenv